+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## $v
+## v170226.32361
 
 - Adding support for Elite variation.
 - Adding support for `async` scripts.